+++ packages/client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Throw an error when an incorrect type is passed to `FixedSizeBinary`
+
 ## 3.1.0 - 2026-09-01
 
 ### Added